Move LogBrowser into separate file
